Labs/Lab2/Python/requirements.txt
networkx
numpy
attrs
jsoncomment
